перенос функции из модуля формы в модуль апи
