+++ js/api.js
@@ -1,5 +1,5 @@
 import {createMarker} from './map.js';
-import {createErrorMessage} from './alerts.js';
+import {createErrorMessage, showSuccessMessage, showErrorMessage} from './alerts.js';
 
 fetch('https://24.javascript.pages.academy/keksobooking/data')
   .then((response) => {
@@ -13,3 +13,28 @@ fetch('https://24.javascript.pages.academy/keksobooking/data')
   .then((similarOffers) => {
     createMarker(similarOffers.slice(0, 10));
   });
+
+
+const setUserFormSubmit = (onSuccess) => {
+  evt.preventDefault();
+
+  const formData = new FormData(evt.target);
+  fetch(
+    'https://24.javascript.pages.academy/keksobooking',
+    {
+      method: 'POST',
+      body: formData,
+      enctype: 'multipart/form-data',
+    },
+    ).then((response) => {
+    if (response.ok) {
+      showSuccessMessage();
+      onSuccess();
+    } else {
+      showErrorMessage();
+    }
+  })
+    .catch(() => {
+      showErrorMessage();
+    });
+});
